Moved filteredByList and updateFilteredByList out of app store and into QueueTable and local state
