+++ client/app/queue/reducers.js
@@ -43,7 +43,6 @@ export const initialState = {
   attorneysOfJudge: [],
   attorneyAppealsLoadingState: {},
   isTaskAssignedToUserSelected: {},
-  filteredByList: {},
   pendingDistribution: null,
   attorneys: {},
   organizationId: null,
@@ -88,12 +87,6 @@ export const workQueueReducer = (state: QueueState = initialState, action: Objec
       },
       amaTasks: {
         $merge: action.payload.amaTasks ? action.payload.amaTasks : {}
-      }
-    });
-  case ACTIONS.UPDATE_FILTERED_BY_LIST:
-    return update(state, {
-      filteredByList: {
-        $set: action.payload.newList
       }
     });
   case ACTIONS.RECEIVE_AMA_TASKS:
